3/ClientPortal.Web/ClientPortal.Web.Tests/Controllers/HomeControllerTest.cs
using ClientPortal.Web.Controllers;
using DirectAgents.Mvc.KendoGridBinder;
using Microsoft.VisualStudio.TestTools.UnitTesting;
using System;
using System.Web.Mvc;

namespace ClientPortal.Web.Tests.Controllers
{
    [TestClass]
    public class HomeControllerTest
    {
        [TestMethod]
        public void OfferSummaryGrid()
        {
            HomeController controller = new HomeController();
            KendoGridRequest kendoGridRequest = new KendoGridRequest
            {
                 
            };
            JsonResult result = controller.OfferSummaryGrid(kendoGridRequest);
            Console.WriteLine(result.Data);
        }

        [TestMethod]
        public void Index()
        {
            // Arrange
            HomeController controller = new HomeController();

            // Act
            ViewResult result = controller.Index() as ViewResult;

            // Assert
            Assert.AreEqual("Modify this template to jump-start your ASP.NET MVC application.", result.ViewBag.Message);
        }

        [TestMethod]
        public void About()
        {
            // Arrange
            HomeController controller = new HomeController();

            // Act
            ViewResult result = controller.About() as ViewResult;

            // Assert
            Assert.IsNotNull(result);
        }

        [TestMethod]
        public void Contact()
        {
            // Arrange
            HomeController controller = new HomeController();

            // Act
            ViewResult result = controller.Contact() as ViewResult;

            // Assert
            Assert.IsNotNull(result);
        }
    }
}
